refactor : 명령어 관련 클래스 수정과 service 패키지 내 클래스들 제거
- status 명령어 추가
- 현재 구현 상황에서 service 클래스로 실행할 필요성을 느끼지 못해 우선 제거
