+++ src/main/java/chess/domain/command/MoveCommand.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.command;
 
 import chess.domain.board.Position;
 import chess.domain.board.Positions;
