--- src/main/java/chess/domain/service/GameEndService.java
@@ -1,8 +1,0 @@
-package chess.domain.service;
-
-public class GameEndService implements GameService {
-	@Override
-	public void run() {
-		System.exit(0);
-	}
-}
