--- src/main/java/chess/domain/service/GameService.java
@@ -1,5 +1,0 @@
-package chess.domain.service;
-
-public interface GameService {
-    void run();
-}
